Created a component to render the Logo and remove duplicate code
